pyproject.toml
[tool.poetry]
name = "building_releases"
version = "0.0.0"
description = ""
authors = ["GlebMauzr <mag@letsnova.ru>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
